requirements.txt
playwright==1.39.0
asyncio==3.4.3
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3 